Move ethereum and hermez types to specific domain namespaces
