+++ src/persistence/index.ts
@@ -1,4 +1,4 @@
-import { HermezTransaction, Exit } from '../domain'
+import { HermezTransaction, Exit } from '../domain/hermez'
 
 export interface HistoryTransactions { transactions: HermezTransaction[], pendingItems: number };
 export interface HistoryExits { exits: Exit[], pendingItems: number };
